Add files via upload
Our data set is from Kaggle and contains the transactions made by credit cards in two days of September 2013 by European cardholders. The data consists of  284807 rows and has 30 features.  Our dataset is highly imbalance. Time variable indicates the amount of time in seconds that has occurred from each transaction to the first transaction and Amount variable indicates the transaction amount. ‘Class’ indicates 1 for fraudulent and 0 for not fraudulent.

In order to detect and combat fraudulent credit card transactions, so that customers are not charged for items they did not purchase. We built a few machine learning  models a baseline logistic regression, Decision Tree, Random Forest, and a Deep Neural Networks.